+++ src/AceCommon.h
@@ -2,6 +2,7 @@
 #define ACE_COMMON_H
 
 #include "ace_common/arithmetic.h"
+#include "ace_common/pstrings.h"
 
 // Version format: "xx.yy.zz" => xxyyzz (without leading 0)
 #define ACE_COMMON_VERSION 10000