+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 * Unreleased
-    * Add `src/ace_common/arithmetic.h` and unit tests.
+    * Add `src/ace_common/arithmetic.h` and unit tests
+    * Add `src/ace_common/pstrings.h` and unit tests